bugfix and refactor
