+++ src/main/java/ure/areas/ULandscaper.java
@@ -10,7 +10,7 @@ import ure.sys.Injector;
 import ure.sys.UCommander;
 import ure.ui.ULight;
 import ure.math.UColor;
-import ure.math.USimplexNoise;
+import ure.math.SimplexNoise;
 import ure.terrain.Stairs;
 import ure.terrain.UTerrain;
 import ure.terrain.UTerrainCzar;
@@ -51,7 +51,7 @@ public abstract class ULandscaper {
     public URandom random;
 
     @JsonIgnore
-    USimplexNoise simplexNoise = new USimplexNoise();
+    SimplexNoise simplexNoise = new SimplexNoise();
 
     protected String floorterrain = "rock";
