+++ src/main/java/ure/sys/UREgame.java
@@ -1,6 +1,6 @@
 package ure.sys;
 
-public interface UREGame {
+public interface UREgame {
 
     public void setupTitleScreen();
